Module framework: reduce input/output dichotomy
There was a lot of repeated code for inputs and outputs, which has now
been reduced where applicable. For example, "addInput(ModuleItem)" and
"addOutput(ModuleItem)" is now simply "addItem(ModuleItem)".

Further, the Module interface now extends Map<String, Object>, which
standardizes the getting and setting of module item values.

This change also fixes a couple of bugs:
  * DefaultModule#addOutput(String, Class<?>) neglected to set the newly
    created DefaultModuleItem's I/O type to ItemIO.OUTPUT.
  * DynamicCommand#getOutput(String) erroneously delegated to
    super.getInput(String) in the case of a non-annotated module item.

Both of the above issues are now moot due to the refactoring.